requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiomysql==0.2.0
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
commonmark==0.9.1
cryptography==44.0.2
cx_Freeze==7.2.10
cx_Logging==3.2.1
dacite==1.9.2
DBUtils==3.1.0
decompyle3==3.9.2
decorator==5.1.1
Deprecated==1.2.18
et_xmlfile==2.0.0
fastapi==0.103.2
filelock==3.17.0
fiscal==0.2.3
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
git-filter-repo==2.47.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.0.0
instaloader==4.14.1
itsdangerous==2.2.0
jaconv==0.4.0
Jinja2==3.1.5
lief==0.16.3
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moviepy==2.1.2
multidict==6.1.0
mutagen==1.47.0
mysql-connector-python==9.2.0
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.7
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pika==1.3.2
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
proglog==0.1.10
progress==1.6
propcache==0.3.0
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pycpfcnpj==1.8
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pykakasi==2.3.0
PyMySQL==1.1.1
PyMysqlPool==1.0.9
PyNFe==0.5.3
pyOpenSSL==25.0.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytube==15.0.0
pytube3==9.6.4
pytz==2025.1
pywin32==308
pywin32-ctypes==0.2.3
pyzmq==26.3.0
qasync==0.27.1
RapidFuzz==3.12.1
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
sentry-sdk==1.45.1
setuptools==75.8.0
shiboken6==6.8.2.1
signxml==4.0.3
six==1.17.0
sniffio==1.3.1
soundcloud-v2==1.6.0
soupsieve==2.6
spark-parser==1.8.9
spotdl==4.2.11
spotify_dl==8.9.0
spotipy==2.25.0
SQLAlchemy==2.0.38
starlette==0.27.0
static_ffmpeg==2.8
syncedlyrics==1.0.1
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uncompyle6==3.9.2
Unidecode==1.3.8
urllib3==1.26.20
uvicorn==0.23.2
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
win10toast==0.9
wrapt==1.17.2
xdis==6.1.1
yarl==1.18.3
yt-dlp==2025.2.19
ytmusicapi==1.10.1
